backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
asgiref==3.11.0
    # via django
billiard==4.2.4
    # via celery
celery==5.6.2
    # via -r requirements.in
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
django==5.2.10
    # via
    #   -r requirements.in
    #   django-ninja
    #   django-redis
django-ninja==1.5.3
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
exceptiongroup==1.3.1
    # via celery
kombu==5.6.2
    # via celery
packaging==25.0
    # via kombu
prompt-toolkit==3.0.52
    # via click-repl
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via django-ninja
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via celery
redis==7.1.0
    # via
    #   -r requirements.in
    #   django-redis
six==1.17.0
    # via python-dateutil
sqlparse==0.5.5
    # via django
typing-extensions==4.15.0
    # via
    #   asgiref
    #   exceptiongroup
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via
    #   django
    #   kombu
    #   tzlocal
tzlocal==5.3.1
    # via celery
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
